domain - infra 엔티티 분리 철회
