--- src/main/java/io/hhplus/concertbook/domain/entity/Concert.java
@@ -1,18 +1,0 @@
-package io.hhplus.concertbook.domain.entity;
-
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.ToString;
-
-@Data
-
-public class Concert {
-
-    private long concertId;
-
-    private String concertName;
-    private long fee;
-
-
-}
