--- src/main/java/io/hhplus/concertbook/domain/entity/ConcertItem.java
@@ -1,20 +1,0 @@
-package io.hhplus.concertbook.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.ToString;
-
-@Data
-
-public class ConcertItem {
-
-    long concertItemId;
-
-    long concertId;
-
-    String concertD;
-    String concertT;
-    int allseats;
-    int availSeats;
-
-}
