+++ src/main/java/io/hhplus/concertbook/domain/entity/PaymentEntity.java
@@ -4,14 +4,21 @@ import jakarta.persistence.*;
 import lombok.Data;
 import lombok.ToString;
 
+import java.sql.Time;
 import java.sql.Timestamp;
 
 @Data
+@Entity
+@Table(name="payment")
+public class PaymentEntity {
 
-public class Payment {
-
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="payment_id")
     long paymentId;
 
+    @ManyToOne
+    @JoinColumn(name = "book_id")
     long bookId;
 
     Timestamp createdAt;
