--- src/main/java/io/hhplus/concertbook/domain/entity/Seat.java
@@ -1,18 +1,0 @@
-package io.hhplus.concertbook.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.ToString;
-
-
-@Data
-
-public class Seat {
-
-    long seatId;
-
-    long concertItemId;
-
-    int seatNo;
-    boolean useYn;
-}
