--- src/main/java/io/hhplus/concertbook/domain/entity/User.java
@@ -1,14 +1,0 @@
-package io.hhplus.concertbook.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.ToString;
-
-@Data
-
-public class User {
-
-    long userId;
-
-    String userName;
-}
