+++ src/main/java/io/hhplus/concertbook/domain/entity/WalletEntity.java
@@ -5,12 +5,19 @@ import lombok.Data;
 import lombok.ToString;
 
 @Data
+@Entity
+@Table(name="wallet")
 
-public class Wallet {
+public class WalletEntity {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="wallet_id")
     long walletId;
 
     long amount; //결제 금액
 
+    @ManyToOne
+    @JoinColumn(name = "user_id")
     long userId;
 }
