+++ src/main/java/io/hhplus/concertbook/domain/repository/PaymentRepo.java
@@ -2,5 +2,5 @@ package io.hhplus.concertbook.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepo {
+public interface PaymentRepo extends JpaRepository {
 }
