+++ src/main/java/io/hhplus/concertbook/domain/repository/UserRepo.java
@@ -1,5 +1,7 @@
 package io.hhplus.concertbook.domain.repository;
 
-public interface UserRepo {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepo extends JpaRepository {
 
 }
