--- src/main/java/io/hhplus/concertbook/infrastructure/mapper/MapperConfig.java
@@ -1,12 +1,0 @@
-package io.hhplus.concertbook.infrastructure.mapper;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-
-public class MapperConfig {
-
-    @Bean
-    public ModelMapper modelMapper(){
-        return new ModelMapper();
-    }
-}
