--- src/main/java/io/hhplus/concertbook/infrastructure/repository/BookJpaRepo.java
@@ -1,6 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookJpaRepo extends JpaRepository {
-}
