--- src/main/java/io/hhplus/concertbook/infrastructure/repository/BookRepoImpl.java
@@ -1,6 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import io.hhplus.concertbook.domain.repository.BookRepo;
-
-public class BookRepoImpl implements BookRepo {
-}
