--- src/main/java/io/hhplus/concertbook/infrastructure/repository/ConcertItemRepoImpl.java
@@ -1,6 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import io.hhplus.concertbook.domain.repository.ConcertItemRepo;
-
-public class ConcertItemRepoImpl implements ConcertItemRepo {
-}
