--- src/main/java/io/hhplus/concertbook/infrastructure/repository/ConcertJpaRepo.java
@@ -1,7 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ConcertJpaRepo extends JpaRepository {
-
-}
