--- src/main/java/io/hhplus/concertbook/infrastructure/repository/ConcertRepoImpl.java
@@ -1,7 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import io.hhplus.concertbook.domain.repository.ConcertRepo;
-
-public class ConcertRepoImpl implements ConcertRepo {
-
-}
