--- src/main/java/io/hhplus/concertbook/infrastructure/repository/PaymentRepoImpl.java
@@ -1,6 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import io.hhplus.concertbook.domain.repository.PaymentRepo;
-
-public class PaymentRepoImpl implements PaymentRepo {
-}
