--- src/main/java/io/hhplus/concertbook/infrastructure/repository/SeatJpaRepo.java
@@ -1,6 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SeatJpaRepo extends JpaRepository {
-}
