--- src/main/java/io/hhplus/concertbook/infrastructure/repository/SeatRepoImpl.java
@@ -1,6 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import io.hhplus.concertbook.domain.repository.SeatRepo;
-
-public class SeatRepoImpl implements SeatRepo {
-}
