--- src/main/java/io/hhplus/concertbook/infrastructure/repository/UserJpaRepo.java
@@ -1,7 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserJpaRepo extends JpaRepository {
-
-}
