--- src/main/java/io/hhplus/concertbook/infrastructure/repository/UserRepoImpl.java
@@ -1,7 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import io.hhplus.concertbook.domain.repository.UserRepo;
-
-public class UserRepoImpl implements UserRepo {
-
-}
