--- src/main/java/io/hhplus/concertbook/infrastructure/repository/WaitTokenRepoImpl.java
@@ -1,6 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import io.hhplus.concertbook.domain.repository.WaitTokenRepo;
-
-public class WaitTokenRepoImpl implements WaitTokenRepo {
-}
