--- src/main/java/io/hhplus/concertbook/infrastructure/repository/WalletRepoImpl.java
@@ -1,6 +1,0 @@
-package io.hhplus.concertbook.infrastructure.repository;
-
-import io.hhplus.concertbook.domain.repository.WalletRepo;
-
-public class WalletRepoImpl implements WalletRepo {
-}
